dom/html/test/test_bug405242.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=405242
-->
<head>
  <title>Test for Bug 405252</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=405242">Mozilla Bug 405242</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 405242 **/
var sel = document.createElement("select");
sel.appendChild(new Option());
sel.appendChild(new Option());
sel.appendChild(new Option());
opt = new Option();
opt.value = 10;
sel.appendChild(opt);
sel.options.remove(0);
sel.options.remove(1000);
sel.options.remove(-1);
is(sel.length, 1, "Unexpected option collection length");
is(sel[0].value, 10, "Unexpected remained option");
</script>
</pre>
</body>
</html>
